reorganize: nav -> comp/nav, useroptions -> comp/nav
