+++ frontend/src/components/navbar/UserOptions.tsx
